Initial import.
Rough Outline of Blackjack rules in rough OOP form.

Writing this to refresh my memory after doing lots of data structures work and less functional programming.

Need to properly program Ace to either be 1 or 11, need to program the actual playing of the game.  Could probably use less variables in the card class.